UI changes Added Park details page - with block placeholder for whatever Stefanie will put in. Added expandable list below to hold park notes and anything else we want to list.  Right now the only top level item is "Park Notes" Added button for adding park note and alert bc ... i was already at it. Button for alert doesn't do anything, it's just for show. Feel free to remove it brad. Added park note layout - includes username, date added (need to fix time zone) and note. Notes cannot be removed, only added Modified options menu to use user's full name
Other changes
Added base class for DBQueryTask to make querying easier and less busy around code.  All queries I used are in queries package.
We now use the email that the user is logged in with (not username bc we get email from google sign in and from email signin) to grab the full user from the db *once* and pass that around between activities and fragments.  This is done in the mainactivity. I combined code that Peter wrote into the UserQueryTask and called that.  The processfinished function is a callback from after the query finishes. See examples in code for adding other stuff like it.

Modified LoginActivity to pass email from google signing to mainactivity to enable getting current user.
Modified registrationactivity to use info from the User object instead of making the call to db there (see above for where I moved the db call to UserQueryTask)